+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The editor dropdown list now won't be wider than the window, even if you have
   very long citations or keywords
 - Removed the leading `#` in the tag cloud
+- Allow tags to be sorted by name or count
 
 ## Under the Hood